quelques methodes du crud de Appareil, Evenement et Watchlist

+++ src/main/java/org/mbds/tpt/maslow/dao/AppareilDao.java
@@ -10,6 +10,6 @@ import javax.transaction.Transactional;
  */
 @Transactional
 @org.springframework.stereotype.Repository
-public interface AppareilDao extends CrudRepository<Appareil, Long> {
+public interface AppareilDao extends CrudRepository<Appareil, Integer> {
 
 }
